feat: Add Android installer implementations
