+++ .editorconfig
@@ -1,0 +1,1 @@
+no-wildcard-imports = disabled
