+++ gradle.properties
@@ -1,4 +1,14 @@
-org.gradle.parallel = true
-org.gradle.caching = true
-kotlin.code.style = official
 version = 1.5.0-dev.2
+
+#Gradle
+org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\="-Xmx2048M"
+org.gradle.caching=true
+org.gradle.configuration-cache=true
+org.gradle.parallel = true
+
+#Kotlin
+kotlin.code.style=official
+
+#Android
+android.useAndroidX=true
+android.nonTransitiveRClass=true
