+++ revanced-library/src/androidMain/aidl/app/revanced/library/IRootService.aidl
@@ -1,0 +1,5 @@
+package aidl.app.revanced.library;
+
+interface IRootService {
+    IBinder getFileSystemService();
+}
