+++ revanced-library/src/commonMain/kotlin/app/revanced/library/installation/installer/NoRootPermissionException.kt
@@ -1,0 +1,3 @@
+package app.revanced.library.installation.installer
+
+class NoRootPermissionException internal constructor() : Exception("No root permission")
