+++ settings.gradle.kts
@@ -1,7 +1,30 @@
-rootProject.name = "revanced-library"
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 buildCache {
     local {
-        isEnabled = !System.getenv().containsKey("CI")
+        isEnabled = "CI" !in System.getenv()
     }
 }
+
+pluginManagement {
+    repositories {
+        google()
+        gradlePluginPortal()
+        mavenCentral()
+    }
+}
+
+@Suppress("UnstableApiUsage")
+dependencyResolutionManagement {
+    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+    repositories {
+        mavenCentral()
+        mavenLocal()
+        maven { url = uri("https://jitpack.io") }
+        google()
+    }
+}
+
+rootProject.name = "revanced-library"
+
+include(":revanced-library")
