--- src/main/kotlin/app/revanced/library/installation/command/LocalCommandRunner.kt
@@ -1,6 +1,0 @@
-package app.revanced.library.installation.command
-
-/**
- * The [CommandRunner] for running commands on the current device.
- */
-abstract class LocalCommandRunner : CommandRunner
